Send 403 if contact does not belong to user

+++ delete.php
@@ -16,10 +16,16 @@ $statement->execute([":id"=>$id]);
 if ($statement->rowCount() == 0) {
   http_response_code(404);
   echo "HTTP 404 NOT FOUND";
+  return;
 }
 
-// $statement = $conn->prepare("DELETE FROM contacts WHERE id = :id");
-// $statement->execute([":id"=>$id]);
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo "HTTP 403 UNAUTHORIZED";
+  return;
+}
 
 $conn->prepare("DELETE FROM contacts WHERE id = :id") -> execute([":id"=>$id]);
